% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bumFDR.R
\name{bumFDR}
\alias{bumFDR}
\title{Fit null and estimate FDR using the BUM model}
\usage{
bumFDR(x, statistic = c("normal", "pvalue"), fit = c("smoothing", "native"),
  a = 1e-06, ...)
}
\description{
Fit null and estimate FDR using the BUM model
}
