% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.HNDModel.R
\name{get.random.HNDscore}
\alias{get.random.HNDscore}
\title{Draws statistics from the half normal dacy (HND) model}
\usage{
get.random.HNDscore(m = 1000, eta0 = 0.8)
}
\description{
Draws statistics from the half normal dacy (HND) model
}
