% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomzScore.R
\name{get.random.zscore}
\alias{get.random.zscore}
\title{Simulate z-scores according to a mixture model}
\usage{
get.random.zscore(d = 200, sigma = 2, eta0 = 0.8, alt.min = 5,
  alt.max = 10)
}
\description{
\deqn{\eta_0 * N(0,\sigma^2) + (1-\eta_0/2) * Unif(alt.min,alt.max)
	+ (1-\eta_0/2)*Unif(-alt.max, alt.min)}
}
\details{
also returns functions of the simulated null and alternative densities and
CDFs as well as FDR and fdr functions
}
