% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingColors.R
\docType{data}
\name{plottingColors}
\alias{plottingColors}
\title{Named vectors of colors used to plot the FDRs, densities etc.}
\format{An object of class \code{character} of length 9.}
\usage{
plottingColors
}
\description{
Colors for plotting
}
\keyword{datasets}
