% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qval.lfdr.pval.R
\name{qval.lfdr.pval}
\alias{qval.lfdr.pval}
\title{Estimate the density and distribution function estimation using logcondens
on  p-values transformed to z-values
constrained such that the known fraction eta0 of null p-values 
is taken into account}
\usage{
qval.lfdr.pval(pv, eta0 = 1)
}
\description{
Estimate the density and distribution function estimation using logcondens
on  p-values transformed to z-values
constrained such that the known fraction eta0 of null p-values 
is taken into account
}
