% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hndFDR.R
\name{run.hnd}
\alias{run.hnd}
\title{Fit null and estimate FDR using the HND model}
\usage{
run.hnd(x, statistic = c("normal", "pvalue"), fit = c("logfdr", "native"),
  ...)
}
\description{
Fit null and estimate FDR using the HND model
}
