% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.cutoff.R
\name{smoothing.cutoff}
\alias{smoothing.cutoff}
\title{### Roxygen based Documentation
estimate tail area based FDR and local fdr  via log-concave densities}
\usage{
smoothing.cutoff(x, statistic = c("normal", "correlation"))
}
\description{
### Roxygen based Documentation
estimate tail area based FDR and local fdr  via log-concave densities
}
